NSInteger

+++ Fat Ninja/MinigameBreakout.mm
@@ -357,7 +357,7 @@ float brokenBlocks;
         evaluation = brokenBlocks/blockCount;
    ballBody->SetType(b2_staticBody);
     //NSLog(@"Win %f", evaluation);
-    int powerDuration;
+    NSInteger powerDuration;
     //abstufung je nachdem wie gut man ein spiel geschafft hat
     if (evaluation == 0) {
         powerDuration = 0;
